Rename RuntimeObjectCreator to FallbackRenderableFactory + clarify two-path model
The class never created runtimes; it produces raw IRenderables for base-type
names. Its real role is the *fallback* path used when the primary registration
path (ElementSaveExtensions.RegisterGueInstantiation) hasn't already supplied
a renderable. The old name actively misled readers about which registry is
the source of truth.

What changed:
- Rename Gum/Wireframe/RuntimeObjectCreator.cs to FallbackRenderableFactory.cs;
  update all <Compile Include> links (MonoGameGum/KniGum/FnaGum/RaylibGum/
  SokolGum/EditorTabPlugin_XNA) and the GumCoreShared.projitems entry that
  FRB1's GumCore csprojs import.
- Keep [Obsolete] RuntimeObjectCreator shim in the same file that forwards
  to FallbackRenderableFactory so external consumers keep compiling with a
  deprecation warning steering them to the new name.
- Update internal call sites (MonoGameGum.Renderables.RenderableCreator,
  Tool EditorTabPlugin MainEditorTabPlugin) and rename the three test files
  (MG/Raylib/Sokol) to match.

Doc additions explaining the two-registry model end to end:
- Extensive XML <remarks> on FallbackRenderableFactory: when this fallback
  fires (only when the primary path leaves RenderableComponent null), why
  it can silently diverge from the primary path (the Apos.Shapes Circle
  example), and that new base types should be added via RegisterGueInstantiation
  rather than extending this switch.
- XML docs on ElementSaveExtensions.CustomCreateGraphicalComponentFunc (the
  wiring point) and RegisterGueInstantiation (the primary path) that
  cross-reference each other and link issue #2915.

No code-path change: behavior is identical to the prior class. The collapse
of paths discussed in #2915 is intentionally out of scope here.

Builds: AllLibraries.sln + GumFull.sln clean. Tests:
FallbackRenderableFactoryTests pass 11/11 on each of MG, Raylib, Sokol.
