+++ GumCoreShared.projitems
@@ -73,7 +73,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Gum\RenderingLibrary\BlendExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Gum\Wireframe\CustomSetPropertyOnRenderable.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Gum\Wireframe\ElementWithState.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Gum\Wireframe\RuntimeObjectCreator.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Gum\Wireframe\FallbackRenderableFactory.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Gum\Wireframe\TimeManager.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ILocalizationService.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MonoGameIntegration\RuntimeLogic.cs" />
