+++ MonoGameGum.Tests/Wireframe/FallbackRenderableFactoryTests.cs
@@ -7,12 +7,12 @@ using Xunit;
 namespace MonoGameGum.Tests.Wireframe;
 
 /// <summary>
-/// Unit tests for <see cref="RuntimeObjectCreator.TryHandleAsBaseType"/>, the factory that
+/// Unit tests for <see cref="FallbackRenderableFactory.TryHandleAsBaseType"/>, the factory that
 /// maps a Gum standard element base type name to its backing <see cref="IRenderable"/>.
 /// This type is compiled into every XNA-like runtime and the Gum tool, so a regression here
 /// breaks both runtime rendering and the editor.
 /// </summary>
-public class RuntimeObjectCreatorTests : BaseTestClass
+public class FallbackRenderableFactoryTests : BaseTestClass
 {
     public override void Dispose()
     {
@@ -24,14 +24,14 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_Circle_ReturnsLineCircle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Circle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Circle", null);
         result.ShouldBeOfType<LineCircle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_ColoredRectangle_ReturnsSolidRectangle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("ColoredRectangle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("ColoredRectangle", null);
         result.ShouldBeOfType<SolidRectangle>();
     }
 
@@ -42,7 +42,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
         // whenever ShowLineRectangles was false (its default), breaking old/XML-error projects.
         GraphicalUiElement.ShowLineRectangles = false;
 
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Component", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Component", null);
 
         result.ShouldNotBeNull();
         result.ShouldBeOfType<InvisibleRenderable>();
@@ -53,7 +53,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     {
         GraphicalUiElement.ShowLineRectangles = true;
 
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Container", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Container", null);
 
         result.ShouldBeOfType<LineRectangle>();
     }
@@ -66,7 +66,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
         // rather than null. A null here leaves the GraphicalUiElement with no contained object.
         GraphicalUiElement.ShowLineRectangles = false;
 
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Container", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Container", null);
 
         result.ShouldNotBeNull();
         result.ShouldBeOfType<InvisibleRenderable>();
@@ -75,35 +75,35 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_NineSlice_ReturnsNineSlice()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("NineSlice", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("NineSlice", null);
         result.ShouldBeOfType<NineSlice>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Polygon_ReturnsLinePolygon()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Polygon", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Polygon", null);
         result.ShouldBeOfType<LinePolygon>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Rectangle_ReturnsLineRectangle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Rectangle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Rectangle", null);
         result.ShouldBeOfType<LineRectangle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Sprite_ReturnsSprite()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Sprite", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Sprite", null);
         result.ShouldBeOfType<Sprite>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Text_ReturnsText()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Text", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Text", null);
         result.ShouldBeOfType<Text>();
     }
 
@@ -112,7 +112,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     {
         // A non-standard name (e.g. a custom component's own name) is expected to fall through;
         // ElementSaveExtensions.CreateGraphicalComponent relies on this null to recurse into base types.
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("SomeCustomComponent", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("SomeCustomComponent", null);
         result.ShouldBeNull();
     }
 }
