+++ MonoGameGum/Renderables/RenderableCreator.cs
@@ -19,7 +19,7 @@ public static class RenderableCreator
 
         IRenderable containedObject = null;
 
-        containedObject = RuntimeObjectCreator.TryHandleAsBaseType(type, systemManagers as SystemManagers);
+        containedObject = FallbackRenderableFactory.TryHandleAsBaseType(type, systemManagers as SystemManagers);
 
 
         // todo - have a custom method...
