+++ Tests/RaylibGum.Tests/Wireframe/FallbackRenderableFactoryTests.cs
@@ -6,16 +6,16 @@ using Shouldly;
 namespace RaylibGum.Tests.Wireframe;
 
 /// <summary>
-/// Unit tests for <see cref="RuntimeObjectCreator.TryHandleAsBaseType"/> as compiled for the
+/// Unit tests for <see cref="FallbackRenderableFactory.TryHandleAsBaseType"/> as compiled for the
 /// Raylib runtime. The same source file backs every runtime via <c>#if</c> fences, but the
 /// Raylib path diverges from the XNA-like runtimes: Container/Component always resolve to an
 /// <see cref="InvisibleRenderable"/> and never honor <see cref="GraphicalUiElement.ShowLineRectangles"/>.
 /// </summary>
-public class RuntimeObjectCreatorTests : BaseTestClass
+public class FallbackRenderableFactoryTests : BaseTestClass
 {
     public override void Dispose()
     {
-        // The XNA-like RuntimeObjectCreator path reads this static; reset it so the
+        // The XNA-like FallbackRenderableFactory path reads this static; reset it so the
         // ShowLineRectangles toggle test cannot leak into other tests.
         GraphicalUiElement.ShowLineRectangles = false;
         base.Dispose();
@@ -24,21 +24,21 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_Circle_ReturnsLineCircle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Circle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Circle", null);
         result.ShouldBeOfType<LineCircle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_ColoredRectangle_ReturnsSolidRectangle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("ColoredRectangle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("ColoredRectangle", null);
         result.ShouldBeOfType<SolidRectangle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Component_ReturnsInvisibleRenderable()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Component", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Component", null);
 
         result.ShouldNotBeNull();
         result.ShouldBeOfType<InvisibleRenderable>();
@@ -47,7 +47,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_Container_ReturnsInvisibleRenderable()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Container", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Container", null);
 
         result.ShouldNotBeNull();
         result.ShouldBeOfType<InvisibleRenderable>();
@@ -61,7 +61,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
         // locks in the platform divergence so it is not "fixed" to match MonoGame by accident.
         GraphicalUiElement.ShowLineRectangles = true;
 
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Container", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Container", null);
 
         result.ShouldBeOfType<InvisibleRenderable>();
     }
@@ -69,35 +69,35 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_NineSlice_ReturnsNineSlice()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("NineSlice", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("NineSlice", null);
         result.ShouldBeOfType<NineSlice>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Polygon_ReturnsLinePolygon()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Polygon", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Polygon", null);
         result.ShouldBeOfType<LinePolygon>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Rectangle_ReturnsLineRectangle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Rectangle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Rectangle", null);
         result.ShouldBeOfType<LineRectangle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Sprite_ReturnsSprite()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Sprite", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Sprite", null);
         result.ShouldBeOfType<Sprite>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Text_ReturnsText()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Text", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Text", null);
         result.ShouldBeOfType<Text>();
     }
 
@@ -106,7 +106,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     {
         // A non-standard name (e.g. a custom component's own name) is expected to fall through;
         // ElementSaveExtensions.CreateGraphicalComponent relies on this null to recurse into base types.
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("SomeCustomComponent", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("SomeCustomComponent", null);
         result.ShouldBeNull();
     }
 }
