+++ Tests/SokolGum.Tests/Wireframe/FallbackRenderableFactoryTests.cs
@@ -6,13 +6,13 @@ using Shouldly;
 namespace SokolGum.Tests.Wireframe;
 
 /// <summary>
-/// Unit tests for <see cref="RuntimeObjectCreator.TryHandleAsBaseType"/> as compiled for the
+/// Unit tests for <see cref="FallbackRenderableFactory.TryHandleAsBaseType"/> as compiled for the
 /// Sokol runtime. The same source file backs every runtime via <c>#if</c> fences; the Sokol
 /// path shares the Raylib branch where it matters (Container/Component always resolve to an
 /// <see cref="InvisibleRenderable"/>) and uses Sokol's own renderable types for everything
 /// else.
 /// </summary>
-public class RuntimeObjectCreatorTests : BaseTestClass
+public class FallbackRenderableFactoryTests : BaseTestClass
 {
     public override void Dispose()
     {
@@ -23,21 +23,21 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_Circle_ReturnsLineCircle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Circle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Circle", null);
         result.ShouldBeOfType<LineCircle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_ColoredRectangle_ReturnsSolidRectangle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("ColoredRectangle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("ColoredRectangle", null);
         result.ShouldBeOfType<SolidRectangle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Component_ReturnsInvisibleRenderable()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Component", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Component", null);
 
         result.ShouldNotBeNull();
         result.ShouldBeOfType<InvisibleRenderable>();
@@ -46,7 +46,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_Container_ReturnsInvisibleRenderable()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Container", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Container", null);
 
         result.ShouldNotBeNull();
         result.ShouldBeOfType<InvisibleRenderable>();
@@ -60,7 +60,7 @@ public class RuntimeObjectCreatorTests : BaseTestClass
         // so it is not "fixed" to match the MonoGame branch by accident.
         GraphicalUiElement.ShowLineRectangles = true;
 
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Container", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Container", null);
 
         result.ShouldBeOfType<InvisibleRenderable>();
     }
@@ -68,42 +68,42 @@ public class RuntimeObjectCreatorTests : BaseTestClass
     [Fact]
     public void TryHandleAsBaseType_NineSlice_ReturnsNineSlice()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("NineSlice", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("NineSlice", null);
         result.ShouldBeOfType<NineSlice>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Polygon_ReturnsLinePolygon()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Polygon", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Polygon", null);
         result.ShouldBeOfType<LinePolygon>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Rectangle_ReturnsLineRectangle()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Rectangle", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Rectangle", null);
         result.ShouldBeOfType<LineRectangle>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Sprite_ReturnsSprite()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Sprite", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Sprite", null);
         result.ShouldBeOfType<Sprite>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_Text_ReturnsText()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("Text", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("Text", null);
         result.ShouldBeOfType<Text>();
     }
 
     [Fact]
     public void TryHandleAsBaseType_UnrecognizedType_ReturnsNull()
     {
-        IRenderable result = RuntimeObjectCreator.TryHandleAsBaseType("SomeCustomComponent", null);
+        IRenderable result = FallbackRenderableFactory.TryHandleAsBaseType("SomeCustomComponent", null);
         result.ShouldBeNull();
     }
 }
