+++ Tool/EditorTabPlugin_XNA/EditorTabPlugin_XNA.csproj
@@ -28,7 +28,7 @@
 		<LangVersion>12.0</LangVersion>
 	</PropertyGroup>
 	<ItemGroup>
-	  <Compile Include="..\..\Gum\Wireframe\RuntimeObjectCreator.cs" Link="Services\RuntimeObjectCreator.cs" />
+	  <Compile Include="..\..\Gum\Wireframe\FallbackRenderableFactory.cs" Link="Services\FallbackRenderableFactory.cs" />
 	</ItemGroup>
